+++ internal/commands/history/history.go
@@ -22,7 +22,7 @@ import (
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 
-	"github.com/fidelity/kconnect/internal/app"
+	"github.com/fidelity/kconnect/pkg/app"
 	"github.com/fidelity/kconnect/pkg/config"
 	"github.com/fidelity/kconnect/pkg/flags"
 	"github.com/fidelity/kconnect/pkg/utils"
@@ -34,7 +34,7 @@ const (
 	longDesc          = `
 Command to allow users to import or export history files.
 
-A common use case would be for one member of a team to generate the history + 
+A common use case would be for one member of a team to generate the history +
 alias config for their teams cluster(s). They could then send this file out to
 the rest of the team, who can then import it. On import, they can set their
 username for all of the history entries.
@@ -87,6 +87,12 @@ func Command() (*cobra.Command, error) {
 		return nil, fmt.Errorf("creating history export command: %w", err)
 	}
 	historyCmd.AddCommand(exportCmd)
+
+	rmCmd, err := rmCommand()
+	if err != nil {
+		return nil, fmt.Errorf("creating history rm command: %w", err)
+	}
+	historyCmd.AddCommand(rmCmd)
 
 	return historyCmd, nil