+++ pkg/app/app.go
@@ -22,10 +22,17 @@ import (
 	"github.com/spf13/pflag"
 	"go.uber.org/zap"
 
-	"github.com/fidelity/kconnect/internal/defaults"
+	"github.com/fidelity/kconnect/pkg/defaults"
 	"github.com/fidelity/kconnect/pkg/history"
+	khttp "github.com/fidelity/kconnect/pkg/http"
 	"github.com/fidelity/kconnect/pkg/prompt"
 	"github.com/fidelity/kconnect/pkg/provider"
+	"github.com/fidelity/kconnect/pkg/provider/discovery"
+)
+
+const (
+	// Name is the name of the application
+	Name = "kconnect"
 )
 
 // App represents the kconnect application and contains the
@@ -33,15 +40,21 @@ import (
 type App struct {
 	historyStore    history.Store
 	configDirectory string
-	selectCluster   SelectClusterFunc
 	sensitiveFlags  map[string]*pflag.Flag
+
+	selectCluster SelectClusterFunc
+	itemSelector  provider.SelectItemFunc
+
+	interactive bool
+	httpClient  khttp.Client
+	logger      *zap.SugaredLogger
 }
 
 // Option represents an option to use with the kcinnect application
 type Option func(*App)
 
 // SelectClusterFunc is a function type that is used to allow selection of a cluster
-type SelectClusterFunc func(discoverOutput *provider.DiscoverOutput) (*provider.Cluster, error)
+type SelectClusterFunc func(discoverOutput *discovery.DiscoverOutput) (*discovery.Cluster, error)
 
 // New creates a new instance of the kconnect app with options
 func New(opts ...Option) *App {
@@ -49,6 +62,10 @@ func New(opts ...Option) *App {
 		configDirectory: defaults.AppDirectory(),
 		selectCluster:   DefaultSelectCluster,
 		sensitiveFlags:  make(map[string]*pflag.Flag),
+		logger:          zap.S().With("app", Name),
+		httpClient:      khttp.NewHTTPClient(),
+		interactive:     true,
+		itemSelector:    provider.DefaultItemSelection,
 	}
 
 	for _, opt := range opts {
@@ -72,10 +89,40 @@ func WithHistoryStore(store history.Store) Option {
 	}
 }
 
+func WithLogger(logger *zap.SugaredLogger) Option {
+	return func(a *App) {
+		a.logger = logger.With("app", Name)
+	}
+}
+
+func WithHTTPClient(client khttp.Client) Option {
+	return func(a *App) {
+		a.httpClient = client
+	}
+}
+
+func WithItemSelectorFunc(itemSelector provider.SelectItemFunc) Option {
+	return func(a *App) {
+		a.itemSelector = itemSelector
+	}
+}
+
+func WithSelectClusterFunc(selectFunc SelectClusterFunc) Option {
+	return func(a *App) {
+		a.selectCluster = selectFunc
+	}
+}
+
+func WithInteractive(interactive bool) Option {
+	return func(a *App) {
+		a.interactive = interactive
+	}
+}
+
 // DefaultSelectCluster is the default cluster selection function. If there is only
 // 1 cluster then it automatically selects it. If there are more than 1 cluster then
 // a selection is displayed and the user must choose one
-func DefaultSelectCluster(discoverOutput *provider.DiscoverOutput) (*provider.Cluster, error) {
+func DefaultSelectCluster(discoverOutput *discovery.DiscoverOutput) (*discovery.Cluster, error) {
 	clusterNameToID := make(map[string]string)
 	options := []string{}
 	for _, cluster := range discoverOutput.Clusters {
@@ -92,4 +139,8 @@ func DefaultSelectCluster(discoverOutput *provider.DiscoverOutput) (*provider.Cl
 	zap.S().Debugw("selected cluster", "id", clusterID)
 
 	return discoverOutput.Clusters[clusterID], nil
+}
+
+func (a *App) SetNonInteractive() {
+	a.interactive = false
 }