+++ pkg/app/config.go
@@ -19,8 +19,8 @@ package app
 import (
 	"fmt"
 
-	"github.com/fidelity/kconnect/internal/defaults"
 	"github.com/fidelity/kconnect/pkg/config"
+	"github.com/fidelity/kconnect/pkg/defaults"
 	"github.com/fidelity/kconnect/pkg/printer"
 )
 
@@ -216,6 +216,21 @@ func AddHistoryExportConfig(cs config.ConfigurationSet) error {
 	}
 	if _, err := cs.String("set", "", "fields to set"); err != nil {
 		return fmt.Errorf("adding set config: %w", err)
+	}
+	return nil
+}
+
+type HistoryRemoveConfig struct {
+	All    bool   `json:"all,omitempty"`
+	Filter string `json:"filter,omitempty"`
+}
+
+func AddHistoryRemoveConfig(cs config.ConfigurationSet) error {
+	if _, err := cs.String("filter", "", "filter to apply to import. Can specify multiple filters by using commas, and supports wilcards (*)"); err != nil {
+		return fmt.Errorf("adding filter config: %w", err)
+	}
+	if _, err := cs.Bool("all", false, "remove all entries"); err != nil {
+		return fmt.Errorf("adding all config: %w", err)
 	}
 	return nil
 }