+++ pkg/app/configure.go
@@ -18,6 +18,7 @@ package app
 
 import (
 	"errors"
+	"context"
 	"fmt"
 	"io"
 	"net/http"
@@ -29,7 +30,6 @@ import (
 
 	"github.com/fidelity/kconnect/pkg/config"
 	"github.com/fidelity/kconnect/pkg/printer"
-	"github.com/fidelity/kconnect/pkg/provider"
 )
 
 // ConfigureInput is the input type for the configure command
@@ -43,7 +43,7 @@ type ConfigureInput struct {
 var ErrNotOKHTTPStatusCode = errors.New("non 200 status code")
 
 // Configuration implements the configure command
-func (a *App) Configuration(ctx *provider.Context, input *ConfigureInput) error {
+func (a *App) Configuration(ctx context.Context, input *ConfigureInput) error {
 	if input.SourceLocation == nil || *input.SourceLocation == "" {
 		return a.printConfiguration(input.Output)
 	}